+++ BenQGuru.eMES.Web.Model/TBLTEMPONWIP.cs
@@ -43,6 +43,6 @@ namespace BenQGuru.eMES.Web.Model
         public string RMABILLCODE { get; set; }
         public Nullable<int> MOSEQ { get; set; }
         public Nullable<decimal> SERIAL { get; set; }
-        public short PROCESSED { get; set; }
+        public int PROCESSED { get; set; }
     }
 }
